requirements.txt
# =========================
# FinRAG-Advisor 依赖
# 使用 uv 管理: uv pip install -r requirements.txt
# =========================

# LLM 框架
langchain>=0.3.15
langchain-community>=0.3.15
langchain-core>=0.3.31
langchain-ollama>=0.2.2
langchain-elasticsearch>=0.3.2
langchain-openai>=0.2.14
langchain-text-splitters>=0.3.5
langgraph>=0.2.66

# 向量数据库
elasticsearch>=8.11.0

# 知识图谱
neo4j>=5.15.0

# LLM 客户端（已切换到 DeepSeek API，仅保留 langchain-ollama 用于 OllamaEmbeddings）
# ollama>=0.4.6  # 已移除，对话模型改用 DeepSeek

# Embedding
sentence-transformers>=3.3.1

# 金融数据
akshare>=1.18.30

# 量化回测
akquant>=0.2.13

# Web 界面
streamlit>=1.28.0
plotly>=5.18.0

# 数据处理
pandas>=2.0.0
numpy>=1.26.0

# 评估框架
ragas>=0.4.3

# 工具
python-dotenv>=1.0.1
PyYAML>=6.0
tqdm>=4.67.0
requests>=2.32.0
pypdf>=5.1.0
